requirements.txt
flask==3.0.3
flask-cors==5.0.0
mysql-connector-python==9.1.0

